Removed NFS configuration lines; stop using NFS for synced directory
Issues with using a NFS share for the Vagrant synced folder caused the
sqlite database to be locked and not modifiable by Rails. This was a
problem for our Linux development machines; it worked OK for Macs.
The solution is to stop using NFS for the synced directory, and just
use the native VirtualBox shared folder facilities.

The advantage of using NFS is that there may be a performance penalty
when using the standard shared folder scheme. It is possible that
this issue could be fixed by following instructions on the following
page:

https://www.vagrantup.com/docs/synced-folders/nfs.html

A similar issue is mentioned in this stackoverflow post:

https://stackoverflow.com/questions/38673701/new-sqlite3-database-is-locked

"One should note that POSIX advisory locking is known to be buggy or
even unimplemented on many NFS implementations ... Your best defense is
to not use SQLite for files on a network filesystem."