change search term pattern match to 'contains', and limit results
Pattern matching had been 'starts with', which was giving poor results since
most titles don't start with the term we're interested in. For example, if
we wanted to lookup 'engineer', we would only get results like 'engineering
manager' or something similar, but we're probably actually interested in
'electrical engineers', 'mechanical engineers', etc.

The down-side is that, since the search is implemented unintelligently,
our results for a general term such as 'engineer' will contain an
overwhelming number of results. There are different ways to deal with
this, one would be to simply limit the results, which would take the
same first n records every time that the search for that term is made;
I opted to take a random sampling of 10 of the result records, so that
each record has an equal chance of appearing in the search results.